Avoid spurious cache miss when query is empty (#30)
The lines after `if not result_set:` are executed (as if there had been a cache miss) if the key was found but the resulting set was empty (`False` in a boolean context). This changes the line for a explicit comparison with `None` (the response if there's a cache miss).